langchain-rag/requirements.txt
langchain
langchain-ollama
langchain-community
langchain-core
fastmcp
faiss-cpu
python-dotenv
asyncio